+++ escalade-business/src/main/java/com/nanoo/business/serviceImpl/TopoServiceImpl.java
@@ -117,7 +117,13 @@ public class TopoServiceImpl implements TopoService {
      */
     @Override
     public void deleteTopo(int topoId){
-        topoRepository.deleteById(topoId);
+        uploadUtil = new UploadUtil();
+        Optional<Topo> topo = topoRepository.findById(topoId);
+    
+        if (topo.isPresent()) {
+            topoRepository.deleteById(topoId);
+            uploadUtil.eraseOldPicture(topo.get().getPicturePath());
+        }
     }
     
     /**
